root/requirements.txt
# Core Python package management
pip
setuptools
wheel

# Jupyter Notebook Support
jupyter
ipykernel

# Data manipulation and analysis
pandas
numpy

# Database connections
vertica_python

# Optimization
gurobipy==11.0

# Logging and configuration
pyyaml

# Cryptography (for encryption.py)
cryptography

# Tabulate (for pretty-printing tables in logs)
tabulate

# Code formatting and linting (recommended for code quality)
black
flake8

# Dependency management (useful for managing project dependencies)
pip-tools